feat: Add support for floating-point and long numbers
Extends the lexer and parser to recognize decimal literals and larger integer values.

Refactors the expression evaluation logic to handle mixed-type arithmetic between Long and Double types, ensuring higher precision across mathematical operations. Division is updated to consistently return floating-point results, and string concatenation is improved to handle various numeric types.

+++ src/test/java/lang/ParserTest.java
@@ -51,8 +51,8 @@ public class ParserTest {
         NumberExpression left = (NumberExpression) bin.getLeft();
         NumberExpression right = (NumberExpression) bin.getRight();
 
-        assertEquals(10, left.getValue());
-        assertEquals(5, right.getValue());
+        assertEquals(10, left.getValue().intValue());
+        assertEquals(5, right.getValue().intValue());
     }
 
     @ParameterizedTest
